aegis/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aegis-mcp-firewall"
version = "0.1.0"
description = "Pre-execution firewall proxy for Model Context Protocol (MCP) tool calls"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "pynacl>=1.5.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
aegis = "aegis.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["aegis*"]
